Move DatastoreIndex objects to common package
This makes them accessible to both appscale-admin and
appscale-datastore.
