+++ AdminServer/appscale/admin/utils.py
@@ -10,6 +10,7 @@ import socket
 import tarfile
 
 from appscale.common.constants import HTTPCodes
+from appscale.common.constants import InvalidConfiguration
 from appscale.common.constants import VERSION_PATH_SEPARATOR
 from appscale.taskqueue import constants as tq_constants
 from appscale.taskqueue.constants import InvalidQueueConfiguration
@@ -19,7 +20,6 @@ from . import constants
 from .constants import (
   CustomHTTPError,
   GO,
-  InvalidConfiguration,
   JAVA,
   SOURCES_DIRECTORY,
   Types,
